refactor: move some shared code to the `normal.jl`
