+++ src/ReactiveMP.jl
@@ -55,6 +55,7 @@ include("approximations/sphericalradial.jl")
 include("approximations/laplace.jl")
 include("approximations/importance.jl")
 include("approximations/optimizers.jl")
+include("approximations/rts.jl")
 include("approximations/linearization.jl")
 include("approximations/unscented.jl")
