+++ src/rules/delta/linearization/in.jl
@@ -33,8 +33,9 @@ end
 
 @rule DeltaFn{f}((:in, k), Marginalisation) (m_out::NormalDistributionsFamily, m_ins::ManyOf{N, NormalDistributionsFamily}, meta::DeltaMeta{M, I}) where {f, N, M <: Linearization, L, I <: NTuple{L, Function}} = begin
     (μs_in, Σs_in) = collectStatistics(m_out, m_ins...)
-    (A, b) = localLinearizationMultiIn(getinverse(meta, k), μs_in)
-    (μ_in, Σ_in, _) = concatenateGaussianMV(μs_in, Σs_in)
+    (A, b)     = localLinearizationMultiIn(getinverse(meta, k), μs_in)
+    joint      = convert(JointNormal, μs_in, Σs_in)
+    μ_in, Σ_in = mean_cov(joint)
     m = A * μ_in + b
     V = A * Σ_in * A'
     return convert(promote_variate_type(variate_form(m), NormalMeanVariance), m, V)
