+++ src/rules/delta/linearization/out.jl
@@ -11,7 +11,7 @@ end
 @rule DeltaFn{f}(:out, Marginalisation) (m_ins::ManyOf{N, NormalDistributionsFamily}, meta::DeltaMeta{M}) where {f, N, M <: Linearization} = begin
     (μs_in, Σs_in) = collectStatistics(m_ins...) # Returns arrays with individual means and covariances
     (A, b) = localLinearizationMultiIn(f, μs_in)
-    (m_fw_in, V_fw_in, _) = concatenateGaussianMV(μs_in, Σs_in)
+    (m_fw_in, V_fw_in) = mean_cov(convert(JointNormal, μs_in, Σs_in))
     m = A * m_fw_in + b
     V = A * V_fw_in * A'
     return convert(promote_variate_type(variate_form(m), NormalMeanVariance), m, V)
@@ -20,7 +20,7 @@ end
 @rule DeltaFn{f}(:out, Marginalisation) (m_out::NormalDistributionsFamily, m_ins::ManyOf{N, NormalDistributionsFamily}, meta::DeltaMeta{M}) where {f, N, M <: Linearization} = begin
     (μs_in, Σs_in) = collectStatistics(m_ins...) # Returns arrays with individual means and covariances
     (A, b) = localLinearizationMultiIn(f, μs_in)
-    (μ_in, Σ_in, _) = concatenateGaussianMV(μs_in, Σs_in)
+    (μ_in, Σ_in) = mean_cov(convert(JointNormal, μs_in, Σs_in))
     m = A * μ_in + b
     V = A * Σ_in * A'
     return convert(promote_variate_type(variate_form(m), NormalMeanVariance), m, V)
